Move specific metadata keys to child class
