+++ backend/model/metadata/src/main/java/co/airy/model/metadata/MetadataKeys.java
@@ -41,9 +41,6 @@ public class MetadataKeys {
     public static class ChannelKeys {
         public static final String NAME = "name";
         public static final String IMAGE_URL = "image_url";
-        public static final String PAGE_ID = "page_id";
-        public static final String PAGE_TOKEN = "page_token";
-        public static final String ACCOUNT_ID = "account_id";
     }
 
     public static class MessageKeys {
