+++ backend/sources/facebook/connector/src/main/java/co/airy/core/sources/facebook/api/model/MetadataKeys.java
@@ -1,0 +1,10 @@
+package co.airy.core.sources.facebook.api.model;
+
+
+public class MetadataKeys extends co.airy.model.metadata.MetadataKeys {
+    public static class ChannelKeys extends co.airy.model.metadata.MetadataKeys.ChannelKeys {
+        public static final String PAGE_ID = "page_id";
+        public static final String PAGE_TOKEN = "page_token";
+        public static final String ACCOUNT_ID = "account_id";
+    }
+}
